Support a 12 word mnemonic as a hsm_secret
This adds support for reading a 12-word BIP39 mnemonic from a plaintext
hsm_secret file. If the file contains a valid mnemonic, hsmtool will
derive the 32-byte seed. There's a few TODOs left through out the code. I think the most important one being that I haven't handled the passphrase case so far. In a BIP39 secret, 512 bits of 64 bytes, the first 32 bytes are for the hsm_seed and the next 32 bytes are for the optional passphrase. We're currently hardcoding the passphrase to an empty string but we will probably want to handle this?

I've also refactoed is_mnemonic_secret to use read_and_validate_mnemonic_file so it's a little tidier.