Refactor hsmtool to remove some redundant function.
gets rid of get_hsm_seed which would prompt the user for a password. We were prompting the user for a password 2-3 times., now only load_hsm_secret prompts the user for a password. The exception is check_hsm which requires you to reenter your password.

Also, added a skeleton of some tests I'd like to add but none of them work just yet.